requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
execnet==1.9.0
executing==1.2.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
outcome==1.2.0
packaging==22.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PySocks==1.7.1
pytest==7.2.1
pytest-rerunfailures==11.1.1
pytest-xdist==3.2.0
python-dateutil==2.8.2
pyzmq==24.0.1
requests==2.28.2
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wcwidth==0.2.5
wsproto==1.2.0
